requirements.txt
Flask==3.1.0
flask_cors==5.0.1
llama_cpp_python==0.2.26
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
torch==2.0.1
transformers==4.30.2
safetensors==0.3.1
pymysql==1.1.0
protobuf==4.24.4
tqdm==4.66.1
requests==2.31.0
pickle5==0.0.12
python-docx==1.0.1
mysql.connector